Create mvp
There are a lot of features in the four year plan I would like to implement, but doing them all would be an immense amount of work. By doing this step by step, we can have a minimum viable product that people can use, and slowly step it up over time.